Remove Nix (Sorry)
At this point, I am having trouble staying up-to-date with all the
changes happening in the upstream Idris 2 repository. Nix is an
additional hurdle, and, unfortunately, I am unable to support Nix as
I have little knowledge about it (I am not using it myself).

On the brighter side, Idris 2 now supports Nix.